Can anyone provide a solution to this questions
Develop different versions of the file web.config, in which to configure different options for user access at the level of subdirectories, individual files, roles. Analyze conflicting role assignments at the level of roles and specific users. Secondly provide variants of the web.config file with an explanation of the assigned user access rules.